+++ logprep/processor/generic_adder/rule.py
@@ -79,12 +79,12 @@ In the following example two files are being used, but only the first existing f
 """
 # pylint: enable=anomalous-backslash-in-string
 
-
+import copy
 from attrs import define, field, validators
 
 from logprep.processor.base.rule import InvalidRuleDefinitionError
 from logprep.processor.field_manager.rule import FieldManagerRule
-from logprep.util.getter import GetterFactory
+from logprep.util.getter import GetterFactory, RefreshableGetter
 
 
 class GenericAdderRule(FieldManagerRule):
@@ -101,7 +101,7 @@ class GenericAdderRule(FieldManagerRule):
         add: dict = field(
             validator=validators.deep_mapping(
                 key_validator=validators.instance_of(str),
-                value_validator=validators.instance_of((str, bool, list)),
+                value_validator=validators.instance_of((str, bool, list, int, float, type(None))),
             ),
             default={},
         )
@@ -143,9 +143,23 @@ class GenericAdderRule(FieldManagerRule):
         first existing file by setting :code:`generic_adder.only_first_existing_file: true`.
         In that case, only one file must exist."""
 
+        _base_add: dict = field(default={}, eq=False)
+        """Stores original add fields for future refreshes of getters"""
+
         # pylint: enable=anomalous-backslash-in-string
 
+        def _refresh_add(self):
+            self.add = copy.deepcopy(self._base_add)
+            self._add_from_path()
+
         def __attrs_post_init__(self):
+            self._base_add = copy.deepcopy(self.add)
+
+            for add_file in self.add_from_file:  # pylint: disable=not-an-iterable
+                getter = GetterFactory.from_string(add_file)
+                if isinstance(getter, RefreshableGetter):
+                    getter.add_callback(self._refresh_add)
+
             if self.add_from_file:
                 self._add_from_path()
 
@@ -159,7 +173,8 @@ class GenericAdderRule(FieldManagerRule):
                     missing_files.append(add_file)
                     continue
                 if isinstance(add_dict, dict) and all(
-                    isinstance(value, str) for value in add_dict.values()
+                    isinstance(value, (str, bool, list, int, float, type(None)))
+                    for value in add_dict.values()
                 ):
                     self.add = {**self.add, **add_dict}
                 else: