+++ logprep/util/configuration.py
@@ -224,7 +224,7 @@ from logprep.util.defaults import (
     ENV_NAME_LOGPREP_CREDENTIALS_FILE,
     MIN_CONFIG_REFRESH_INTERVAL,
 )
-from logprep.util.getter import GetterFactory, GetterNotFoundError
+from logprep.util.getter import GetterFactory, GetterNotFoundError, RefreshableGetterError
 from logprep.util.rule_loader import RuleLoader
 
 logger = logging.getLogger("Config")
@@ -629,10 +629,13 @@ class Configuration:
 
     _config_failure: bool = field(default=False, repr=False, eq=False, init=False)
 
+    _current_variable_values: list = field(default=[], repr=False, eq=False, init=False)
+
     _unserializable_fields = (
         "_getter",
         "_configs",
         "_config_failure",
+        "_current_variable_values",
         "_scheduler",
         "_metrics",
         "_unserializable_fields",
@@ -722,10 +725,7 @@ class Configuration:
         """
         try:
             config_getter = GetterFactory.from_string(config_path)
-            try:
-                config_dict = config_getter.get_json()
-            except (json.JSONDecodeError, ValueError):
-                config_dict = config_getter.get_yaml()
+            config_dict = config_getter.get_dict()
             config = Configuration(**(config_dict | {"getter": config_getter}))
         except TypeError as error:
             raise InvalidConfigurationError(
@@ -786,6 +786,8 @@ class Configuration:
             errors = [*errors, *error.errors]
         if errors:
             raise InvalidConfigurationErrors(errors)
+        if configuration.config_refresh_interval:
+            configuration._init_current_variable_values()
         return configuration
 
     def as_dict(self) -> dict:
@@ -842,12 +844,14 @@ class Configuration:
             if self._config_failure:
                 logger.info("Config refresh recovered from failing source")
             self._config_failure = False
-            if new_config == self:
+            new_variable_values = self._get_variable_values(new_config)
+            if new_config == self and self._current_variable_values == new_variable_values:
                 logger.info(
                     "Configuration version didn't change. Continue running with current version."
                 )
                 self._set_config_refresh_interval(new_config.config_refresh_interval)
                 return
+            self._current_variable_values = new_variable_values
             if new_config.config_refresh_interval is None:
                 new_config.config_refresh_interval = self.config_refresh_interval
             self._configs = new_config._configs  # pylint: disable=protected-access
@@ -870,6 +874,29 @@ class Configuration:
         if errors:
             logger.error("Failed to reload configuration: %s", errors)
             self._metrics.number_of_config_refresh_failures += 1
+
+    def _init_current_variable_values(self):
+        self._current_variable_values = Configuration._get_variable_values(self)
+
+    @staticmethod
+    def _get_variable_values(config_dict: "Configuration") -> list:
+        def get_variable_values(value, variable_values: list | None = None):
+            if variable_values is None:
+                variable_values = []
+
+            if isinstance(value, dict):
+                for val in value.values():
+                    get_variable_values(val, variable_values)
+            elif isinstance(value, (list, tuple, set)):
+                for val in value:
+                    get_variable_values(val, variable_values)
+            elif isinstance(value, str):
+                if value.startswith(("http://", "https://", "file://")):
+                    variable_values.append(GetterFactory.from_string(value).get_json())
+
+        variable_values_res: list = []
+        get_variable_values(config_dict.as_dict(), variable_values_res)
+        return variable_values_res
 
     def _set_config_refresh_interval(self, config_refresh_interval: int | None) -> None:
         if config_refresh_interval is None:
@@ -938,7 +965,13 @@ class Configuration:
             try:
                 processor_definition_with_rules = self._load_rule_definitions(processor_definition)
                 pipeline_with_loaded_rules.append(processor_definition_with_rules)
-            except (FactoryError, TypeError, ValueError, InvalidRuleDefinitionError) as error:
+            except (
+                FactoryError,
+                TypeError,
+                ValueError,
+                InvalidRuleDefinitionError,
+                RefreshableGetterError,
+            ) as error:
                 errors.append(error)
         if errors:
             raise InvalidConfigurationErrors(errors)
@@ -1004,7 +1037,13 @@ class Configuration:
                 processor = Factory.create(deepcopy(processor_config))
                 processor.setup()
                 self._verify_rules(processor)
-            except (FactoryError, TypeError, ValueError, InvalidRuleDefinitionError) as error:
+            except (
+                FactoryError,
+                TypeError,
+                ValueError,
+                InvalidRuleDefinitionError,
+                RefreshableGetterError,
+            ) as error:
                 errors.append(error)
             except FileNotFoundError as error:
                 errors.append(InvalidConfigurationError(f"File not found: {error.filename}"))