+++ tests/unit/processor/base.py
@@ -10,7 +10,6 @@ from pathlib import Path
 from unittest import mock
 
 import pytest
-import requests
 import responses
 from attrs import asdict
 from ruamel.yaml import YAML
@@ -25,6 +24,7 @@ from logprep.processor.base.exceptions import (
 )
 from logprep.processor.base.rule import Rule
 from logprep.util.defaults import RULE_FILE_EXTENSIONS
+from logprep.util.getter import HttpGetter, RefreshableGetterError
 from tests.unit.component.base import BaseComponentTestCase
 
 yaml = YAML(typ="safe", pure=True)
@@ -234,11 +234,12 @@ class BaseProcessorTestCase(BaseComponentTestCase):
         assert processor._rule_tree.tree_config.priority_dict == tree_config.get("priority_dict")
 
     @responses.activate
-    def test_raises_http_error(self):
+    def test_raises_http_error_raises_getter_error(self):
         config = deepcopy(self.CONFIG)
         config.update({"tree_config": "http://does.not.matter.bla/tree_config.yml"})
         responses.add(responses.GET, "http://does.not.matter.bla/tree_config.yml", status=404)
-        with pytest.raises(requests.HTTPError):
+        HttpGetter._shared.clear()
+        with pytest.raises(RefreshableGetterError, match="404"):
             Factory.create({"test instance": config})
 
     @pytest.mark.parametrize(