+++ tests/unit/test_runner.py
@@ -106,7 +106,7 @@ class TestRunner:
             with pytest.raises(SystemExit, match=str(EXITCODES.PIPELINE_ERROR.value)):
                 runner.start()
 
-    def test_runner_calls_reload_on_config_change(
+    def test_runner_calls_reload_on_config_version_change(
         self, runner: Runner, configuration: Configuration
     ):
         with mock.patch.object(runner, "_manager") as mock_manager: